requirements.txt
streamlit
tensorflow
pillow
plotly
streamlit-extras
